+++ SF3Lib/Models/Structs/Shared/Color.cs
@@ -15,7 +15,7 @@ namespace SF3.Models.Structs.Shared {
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_colorABGR1555Addr), displayName: "Color (ABGR1555)", displayOrder: 0, displayFormat: "X4")]
         public ushort ColorABGR1555 {
-            get => (ushort) Data.GetUInt16(_colorABGR1555Addr);
+            get => Data.GetUInt16(_colorABGR1555Addr);
             set => Data.SetUInt16(_colorABGR1555Addr, value);
         }
